# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For DH Mobile ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product-specifications": {
            "get": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Lấy danh sách thông số kỹ thuật sản phẩm",
                "operationId": "0ea66054f91cfd32da23b017c58c87fa",
                "responses": {
                    "200": {
                        "description": "Thành công"
                    }
                }
            },
            "post": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Thêm thông số kỹ thuật sản phẩm mới",
                "operationId": "2c61b467d76763faf6db73f79702afc0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "spec_name",
                                    "spec_value"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "spec_name": {
                                        "type": "string"
                                    },
                                    "spec_value": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    }
                }
            }
        },
        "/api/product-specifications/{id}": {
            "get": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Lấy thông số kỹ thuật sản phẩm theo id",
                "operationId": "91c9ec0b7ccd4ae05d990884fc5f91a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Cập nhật thông số kỹ thuật sản phẩm theo id",
                "operationId": "4ffe5c3a155d3fc29dc861402bdf3622",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "spec_name",
                                    "spec_value"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "spec_name": {
                                        "type": "string"
                                    },
                                    "spec_value": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Xóa mềm thông số kỹ thuật sản phẩm theo id",
                "operationId": "93e9a04f17df25e5ba8284f0f91c1429",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
        "/api/product-specifications/restore/{id}": {
            "post": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Khôi phục thông số kỹ thuật đã xóa mềm",
                "operationId": "c0a7bf6de759f06ed7b66d2923eb63bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Khôi phục thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
        "/api/product-specifications/force-delete/{id}": {
            "delete": {
                "tags": [
                    "ProductSpecifications"
                ],
                "summary": "Xóa vĩnh viễn thông số kỹ thuật đã xóa mềm",
                "operationId": "5bf49c83e02131e3f34c54b52ef0a3fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa vĩnh viễn thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}